test(freshness): live triage journey + float8 quality_score cast (GH #34/#35)
Live end-to-end journey for the write-half against the dev graph DB: seed an
ambiguous crystallization → triage surfaces it → keep (success+1) → needs-work
(failure+1) → numeric outcome rejected → retire (superseded_by=id) → leaves
triage → reinstate (override). 7/7 green.

The journey caught that pg returns the numeric quality_score as a string; cast
it to float8 in getMetabolismTriage / metabolizeByEvidence so the runtime matches
the MetabolismRow/TriageRow `number` contract (the console page already coerced).

+++ packages/workers/src/base/memory-repository.ts
@@ -408,7 +408,7 @@ export class PoolMemoryRepository implements MemoryRepository {
          AND (success_count + failure_count) >= $1
          AND ((success_count + 1.0) / (success_count + failure_count + 1.0)) <= $2
        RETURNING id, success_count, failure_count,
-                 ((success_count + 1.0) / (success_count + failure_count + 1.0)) AS quality_score`,
+                 (((success_count + 1.0) / (success_count + failure_count + 1.0)))::float8 AS quality_score`,
       [bands.nMin, bands.qualityBad],
     );
     return rows;
@@ -425,7 +425,7 @@ export class PoolMemoryRepository implements MemoryRepository {
     // meaningful success-rate, so they alone warrant a human decision.
     const { rows } = await this.pool.query<TriageRow>(
       `SELECT id, content, intent_description, success_count, failure_count,
-              ((success_count + 1.0) / (success_count + failure_count + 1.0)) AS quality_score,
+              (((success_count + 1.0) / (success_count + failure_count + 1.0)))::float8 AS quality_score,
               injection_count
        FROM procedural_memory
        WHERE is_anti_pattern = FALSE
